Improved the command bus builder javadocs
This is probably the third documentation refactoring iteration on this class.

These documentation improvements will help the clients using this library to understand better the workings of the instantiation of the bus, and provide knowledge about how to instantiate the bus according to their needs (i.e. on creating a bus bean).

Changelog:
- Added reasons for the existence of this builder class
- Documented the variables meaning
- Improved documentation semantics